requirements.txt
smolagents==1.1.0
transformers==4.46.3
accelerate>=0.26.0
huggingface-hub>=0.25.0,<1.0
trulens-core
trulens-providers-huggingface
openinference-instrumentation-smolagents
gradio>=5.0.0
pandas
duckduckgo-search
matplotlib
seaborn
opencv-python-headless
pdfplumber